reverted to correct logic
I should have stuck with the original logic instead of accepting some proposed changes. I should have explained why the logic was the way I wrote it and also thought thru what was being proposed as an alternative before accepting the proposed change. This will be the way going forward.