build: fetch `jenkins.war` in a separate part
The current build process invokes `curl` as part of an `override-build` scriptlet. As a general rule, external artifacts should be pulled in the `PULL` stage.

In general this hasn't been a problem, but on slower runners such as the emulated `riscv` builders currently used by Launchpad, by the time the `override-build` scriptlet is executed, the proxy token has sometimes expired.

This commit is a minor refactor to fetch `jenkins.war` as a separate part, where the actual fetch will happen in the `PULL` phase of the build.

See [here](https://github.com/snapcrafters/jenkins/actions/runs/16468141666/job/46550507055) for example build failures.